Use unsigned char for lsd input
By running valgrind --tool=massif ./lsd_cmd -C -P pink.eps pink_bg.pgm /dev/null

The peak heap usage reduces from 1175424 to 938824.

That is not a big difference.  But the aim of this commit is not only to reduce
memory usage.  It is because 256bit (unsigned char) grayscale image is more
common than a double array.  pgm image is an example.